4337 signature flow (#64)
Fixes #59

This PR documents the research around ERC-4337 flows, in particular
including signature validation flow.

Most of the rationale for decisions captured in the specification itself
(well, should be in principle 😅). Some of the working assumptions when
implementing this specification are:
1. We **want** 4337 to be implemented over the Safe{Core} Protocol; that
being said, requiring accounts to implement 4337 natively is a viable
alternative (one that ERC-6900 takes for example)
2. We do not want to require a staked paymaster for user operations over
the Safe{Core} Protocol.